requirements.txt
absl-py==0.10.0
asgiref==3.2.10
astor==0.8.1
Babel==2.3.4
beautifulsoup4==4.9.1
bleach==1.5.0
certifi==2020.6.20
chardet==3.0.4
decorator==4.0.10
Django==3.1.2
django-simple-menu==1.2.1
docutils==0.12
ebaysdk==2.1.5
feedparser==5.2.1
gast==0.4.0
greenlet==0.4.16
grpcio==1.32.0
html2text==2016.9.19
html5lib==0.9999999
idna==2.7
importlib-metadata==2.0.0
imutils==0.5.3
Jinja2==2.10.1
lxml==4.5.2
Mako==1.0.4
Markdown==3.2.2
MarkupSafe==0.23
mock==2.0.0
newsapi-python==0.2.6
num2words==0.5.6
numpy==1.19.2
odoo-autodiscover==2.0.0
ofxparse==0.16
opencv-contrib-python-headless==4.4.0.44
passlib==1.6.5
pbr==5.5.0
Pillow==7.2.0
protobuf==3.13.0
psutil==5.7.2
psycopg2==2.8.5
pydot==1.2.3
pyparsing==2.1.10
PyPDF2==1.26.0
pypiwin32==223
pyserial==3.1.1
PySimpleGUI==4.29.0
python-codicefiscale==0.3.7
python-dateutil==2.5.3
python-slugify==4.0.1
python-stdnum==1.14
pytz==2016.7
pyusb==1.0.0
pywin32==228
qrcode==5.3
reportlab==3.5.49
requests==2.20.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
suds-jurko==0.6
tensorboard==1.8.0
tensorflow==1.8.0
termcolor==1.1.0
text-unidecode==1.3
urllib3==1.24.3
vatnumber==1.2
vobject==0.9.3
Werkzeug==0.16.0
wrapt==1.12.1
xlrd==1.0.0
XlsxWriter==0.9.3
xlwt==1.3.0
zipp==3.2.0
